Check tc_lib.add_tc_policy_class input parameters
All input rate parameters (ceil, burst and rate) are mandatory in
Pyroute "tc" class [1]. "tc_lib.add_tc_policy" should provide them.

max_kbps: Now "max_kbps" is mandatory.

burst_kb: In case this parameter does not exist, the library uses the
method previously defined in Neutron, consisting of asigning the 80%
of "max_kbps".

min_kbps: In Pyroute2 [2], the value of "burst" depends on the "rate"
value. If the "rate" value is too small, the "burst" value will exceed
the 2^32 limit defined in the iproute2 structure. In order to avoid
the error described in c#4 of the related bug, the method checks if
the provided "rate" value is greater than the minimum possible value.
If not, the minimum value is set and a warning message is logged.

[1]https://github.com/openstack/neutron/blob/fb2453c999f1757cc9085f4533bb2fd56d6b2245/neutron/privileged/agent/linux/tc_lib.py#L107-L108
[2]https://github.com/svinota/pyroute2/blob/943502a95cbbbac48218d2bde9d92af462682f96/pyroute2/netlink/rtnl/tcmsg/sched_htb.py#L88-L91

Change-Id: If60f4ea9793b0872e348a6319f2c24127e93f540
Related-Bug: #1826565
(cherry picked from commit 1e69279817b4fe978e905987db90dfca9b636b0b)