Mark TestVirtualPorts tests as unstable
Based on debug performed in [1] the TestVirtualPorts
are unstable. At lease one of tests per 20
retries of functional job fail in various
places.
We don't think it is worth to spend more
time on debugging those tests, because there is
a possible easier solution: set virtual port
type based on DEVICE_TYPE, and do not rely
on allowed address pairs.

[1] https://review.opendev.org/#/c/712888/

Related-Bug: #1865453

Change-Id: Iffef8ceb9ca458820a7b3e604a53281f7e93b810