[ovn] Stricter matching on metadata port binding event
Previously the Port Binding event spawned haproxy process even if the
change wasn't done in chassis column, it attempted to spawn haproxy on
all port bindings changes.

This patch spawns haproxy only if new chassis is the one managed by the
agent and old chassis was empty. Similarly it destroys haproxy if new
chassis is empty and old chassis was the one managed by the agent.

Closes-bug: #1868125

Co-Authored-By: Daniel Alvarez <dalvarez@redhat.com>
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>
Change-Id: I5b87726eafa71d717ae22f48d1c9c6343b680c7f