Use rally-openstack 1.7.0 for stable/rocky
Master branch now has requirements incompatible with stable branches
constraints [0], so this switches to last tag compatible.

Longer term fix will be to run rally in a virtualenv, but this is lower
priority than other tasks. This can be revisited at the time, though for
older branches we will probably stick with pinned version

[0] https://opendev.org/openstack/rally-openstack/commit/5776e015f16321fdb489ebb482d984ccd3cd2903

Change-Id: I9ec8331bae54d191955a843f9d3f8c5537dc37f6
Closes-Bug: #1868691