[L3 HA] Add "no_track" option to VIPs in keepalived config
Patch [1] introduced new mechanism which only brings UP interfaces
on master node of HA router. It works fine with keepalived 1.x
but it is broken when keepalived 2.x was used (e.g. on Centos 8) as
in this new version of keepalived by default all interfaces of VIPs
and routes are tracked, and if one of them is DOWN, keepalived is
going to FAULT state. Because of that router will never be
transitioned to MASTER on any node.

This patch fixes it by adding "no_track" option to all VIPs
and routes in keepalived's config file.

This "no_track" option isn't added to ha interface so this one
is still tracked by keepalived.

[1] https://review.opendev.org/#/c/707406/

Closes-bug: #1874211

Change-Id: Ic16cf83fe1d1576d91047adb2d4f9e07d57185b6
(cherry picked from commit dc9084a8ec9db08d7ea947e0b73581b748be5819)