[OVN][metadata] Adding ERROR trace upon unexpected data
When a metadata request arrives to the metadata agent, OVN
will need to figure out which port it corresponds to and it'll do
based on the network ID and IP address which should be a unique pair.

If it's not unique or it doesn't exist, then there is an error
and an ERROR trace is logged.

Change-Id: Id83c2899a03af98e32b0be7ba6ad661e646f07bc
Related-Bug: 1874733
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit 591adfee971906dc86e9357903f164d01657ed65)