[ovn] Add PortBinding delete event for external ports
The ports are unbound from chassis before they get deleted after VM with
normal VIF ports is deleted. In case of SRIOV, we do not unbound the
external SRIOV port but we delete it right away with chassis column set.

This patch adds DELETE event type support to the event that tears down
the metadata datapath. Also a warning is added if bound port that is not
external is being removed.

Change-Id: I17b71933b8c3a7609fcc5027c246f5b03d493ee2
Closes-Bug: #1875865
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>