Optimize subnet retrieval during CIDR validation
If IP overlapping is not allowed, when a new subnet is created, the
assigned CIDR is compared to any other subnet CIDR, searching for
overlapped IP addresses.

To retrieve those subnet CIDRs, there is no need to retrieve the
entire subnet OVO (including all synthetic and lazy fields),  but
only the "id" and the "cidr" fields from the "subnet" table in the
database.

In a development machine, with 500 subnets created, those are the
average times executing both methods:
- Full subnet OVO retrieval ("get_subnets"): 0.9 secs
- "get_subnet_cidrs":                        0.025 secs

Change-Id: I804cf08cbaeec544fbd9645af64d713499052257