Default OVN_BRANCH and OVS_BRANCH to latest stable
If other openstack project uses Neutron with OVN
(as enable_plugin ... neutron) and doesn't specify
OVN_BRANCH and OVS_BRANCH variables in configuration
it will use master branch of those two repositories.

Master branch servers for OVN testing and aligning OVN
Neutron driver with new features.

For now we don't have voting OVN master CI job, to ensure that
everything is fine with it, so we should default OVN for other
projects to the latest stable version that we know works properly.

Change-Id: I6e36ccb8cf4846a1eb11619e7c339a0fc193c4f6