Remove unneeded DB register retrieval and refresh in network update
The first DB operation done during a network update is a network DB
register retrieval. This information is stored in memory during
the DB transaction context and it's updated with any change done
to the DB register and the relationships; e.g., the network DB model
field "qos_policy_binding" is a backref relationship with a
"qos_network_policy_bindings" DB register. When a
"qos_network_policy_bindings" register is created, updated or
deleted, the network DB object, holding the relationship, will be
updated too.

There is no need therefore to retrieve again the DB object and
force a session update. The DB object contains the updated
information.

Trivial-Fix

Change-Id: I0c10b824c0f919607bb6ea31c5533ce7bd65df42