ovn: Fix use_new_ovn_repository
Previously, the function decided if the repository is after the split
based on the OVN branch name. This didn't work when a commit hash or other
branch name was used.

This patch changes the function to look at the tree and decide if the given
commit that split the OVN from OVS is present in the given OVS branch.

Change-Id: I89e1f450e4db4c9554e368d75155bec137074ba6