Bump oslo.log to version 4.1.2
Versions of oslo.log >= 4.1.2 require a second argument when
initializing the OSJournalHandler class. While I've started
a review to fix that regression [0], bump oslo.log to a later
version that requires this new flag and pass it to fix the
gate.

Also had to blacklist the OVN tempest IPv6 hotplug test
since it is failing too much to pass the check jobs.

[0] https://review.opendev.org/732457

Change-Id: Ic9bbb43aa832ad6cc45d57328b40afe4468ddfca
Related-bug: #1871840
Related-bug: #1881558