[OVN] Don't set virtual port type on ports with similar addresses
Because of broken check why one port has similar address to the other
one, the first one by mistake could be set as virtual.
Example:
 port_1: 10.11.1.100
 port_2: 10.11.1.10

On create of port_2 it is set as 'virtual', but shouldn't.

This patch fixes that bug by using common function
utils.get_ovn_port_addresses().

Change-Id: Ia4b986146c77edf0616015380359e37233df80fc
Closes-Bug: #1881759