[OVN] Override notify_nova config in neutron-ovn-db-sync-util
If either ``notify_nova_on_port_status_changes`` or
``notify_nova_on_port_data_changes`` is set in the neutron
configuration provided to the ``neutron-ovn-db-sync-util`` tool it
will stop with a Traceback.

We are already overriding other parts of the ML2 Notification code
as we do not want it executing while syncing, so let's override
these configuration options if set as well.

Change-Id: I62765a14d376664ccc2ca3009bb9e970c9f25c03
Closes-Bug: #1882020