Make neutron functional uwsgi jobs usable by other projects.
neutron-fullstack-with-uwsgi and neutron-functional-with-uwsgi
jobs are run in devstack gate[1] also. On devstack side these jobs
are broken when migrated to zuulv3 and start using neutron defined
tox env. neutron gate is no issue as job find the tox env in neutron's
tox.ini but devstack does not.

We need to define the working dir to neutron so that any where
they are run they always look for the neutron's tox.ini for used
tox env.

Failure: https://review.opendev.org/#/c/577779/

[1] https://opendev.org/openstack/devstack/src/commit/c3b58f5335fb3545bf49b5d542ef9ed702de43c8/.zuul.yaml#L717

Change-Id: I0cadebb45864a83dfba7b54c9f161d11611164db