Fix a typo in the OVN manual install guide
Using a colon is not correct in this context and results in service
startup errors like this:

AttributeError: module 'neutron.services.ovn_l3'
has no attribute 'plugin:OVNL3RouterPlugin'

It could be that this hasn't been modified after copying from setup.cfg
where a colon is used correctly.

Moreover, it would be better to use entrypoint names, so this change
changes that for the L3 service plugin as well as the core plugin.

Change-Id: Id51704a50f42b65049e535aaeb9dc6847a3737bb