Use network.external DB model parameter when creating a floating IP
When a floating IP is being created, the network provided should be
external. Instead of quering the DB to find out if the
"externalnetworks" DB register exists, the "network" register is
retrieved and the "external" parameter is used (loaded using a
back reference relationship). This will avoid one DB access.

Change-Id: Iead245da166ee2ae691227bb18ae377fe0af4c04
Related-Bug: #1880969