Clean up some of the OVN agent API methods
Adds ControllerAgent and MetadataAgent classes to organize some of
the OVN Agent API code.

Note: This is being backported because the change to fix the bug #1892477
relies on these modifications. As #1892477 is a high priority bug I am
backporting this work as part of effort to backport the fix for that
issue.

Partial-Bug: #1892477
Change-Id: I9224edab2d27184ed96be1cdea2bd865aadff065
(cherry picked from commit 6723f4a485e2c60e45f45732c0881c126de06608)